requirements.txt
apsw==3.50.4.0
arrow==1.3.0
asciidoc==10.2.1
attrs==25.3.0
beautifulsoup4==4.14.2
beniget==0.4.2.post1
cachecontrol==0.14.3
cairocffi==1.7.1
certifi==3024.7.22
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
click-plugins==1.1.1
colorama==0.4.6
contourpy==1.3.3
cppy==1.3.1
css-parser==1.0.10
cycler==0.12.1
cython==3.1.2
dependency-groups==1.3.1
deprecated==1.2.18
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring-to-markdown==0.17
docutils==0.22.2
editables==0.5
feedparser==6.0.12
flit-core==3.12.0
fonttools==4.60.1
fqdn==1.5.1
gast==0.6.0
gdbus-codegen==2.84.4
gemato==20.6
hatch-fancy-pypi-readme==25.1.0
hatch-vcs==0.5.0
hatchling==1.27.0
html2text==2025.4.15
html5-parser==0.4.12
html5lib==1.1
idna==3.10
ifaddr==0.2.0
importlib-metadata==8.7.0
iniconfig==2.1.0
installer==0.7.0
isoduration==20.11.0
jaconv==0.4.0
jaraco-collections==5.2.1
jaraco-context==6.0.1
jaraco-functools==4.3.0
jaraco-text==4.0.0
jedi==0.19.2
jeepney==0.9.0
jinja2==3.1.6
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
latex2pydata==0.4.1
latexminted==0.3.1
latexrestricted==0.6.2
lazy-object-proxy==1.12.0
linkify-it-py==2.0.3
lxml==6.0.2
lxml-html-clean==0.4.2
mako==1.3.10
markdown==3.9
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.3
maturin==1.9.4
mdurl==0.1.2
mechanize==0.4.10
meson==1.7.2
meson-python==0.18.0
more-itertools==10.8.0
msgpack==1.1.1
netifaces==0.11.0
nicotine-plus==3.3.10
numpy==2.3.1
olefile==0.47
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pdm-backend==2.4.5
pillow==11.3.0
pip==25.2
pkgcheck==0.10.37
pkgcore==0.12.30
pkgdev==0.2.12
platformdirs==4.4.0
pluggy==1.6.0
ply==3.11
pooch==1.8.2
portage==3.0.69
psutil==7.1.0
pybind11==3.0.0
pycairo==1.28.0
pychm==0.8.6
pycparser==2.23
pycryptodome==3.23.0
pygments==2.19.2
pygobject==3.50.1
pykakasi==2.3.0
pyopengl==3.1.9
pyparsing==3.2.5
pyproject-hooks==1.2.0
pyproject-metadata==0.9.1
pyqt-builder==1.18.2
pyqt5==5.15.11
pyqt5-sip==12.17.1
pyqt6==6.9.1
pyqt6-sip==13.10.2
pyqt6-webengine==6.9.0
pyqtgraph==0.13.7
pysocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.13.1
pythran==0.18.0
pytz==2025.2
pyyaml==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
resolvelib==1.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rich==14.1.0
rpds-py==0.27.1
scipy==1.16.2
setuptools==80.9.0
setuptools-scm==9.2.0
sgmllib3k==1.0.0
sip==6.12.0
six==1.17.0
snakeoil==0.10.11
soupsieve==2.8
strict-rfc3339==0.7
tomli-w==1.2.0
tree-sitter==0.25.1
tree-sitter-bash==0.25.0
trove-classifiers==2025.9.11.17
truststore==0.10.4
typing-extensions==4.15.0
uc-micro-py==1.0.3
ujson==5.11.0
uri-template==1.3.0
urllib3==2.5.0
webcolors==24.11.1
webencodings==0.5.1
wheel==0.45.1
wrapt==1.17.3
xcffib==1.10.0
xxhash==3.5.0
yt-dlp==2025.10.22
zeroconf==0.147.2
zipp==3.23.0
